# resources-service/internal/controller/http/v1/recursos_controller.go
package v1

import (
	"errors"
	"fmt"
	"math"
	"net/http"
	"strconv"
	"strings"

	common_response "github.com/evrone/go-clean-template/internal/controller/http/common/response"
	"github.com/evrone/go-clean-template/internal/controller/http/middleware"
	"github.com/evrone/go-clean-template/internal/controller/http/v1/response"

	"github.com/evrone/go-clean-template/internal/controller/http/v1/request"
	"github.com/evrone/go-clean-template/internal/usecase/common/apperror"
	errorlogports "github.com/evrone/go-clean-template/internal/usecase/errorlog/ports"
	recursosports "github.com/evrone/go-clean-template/internal/usecase/recursos/ports"
	"github.com/evrone/go-clean-template/internal/usecase/recursos/ports/DTOs"
	"github.com/evrone/go-clean-template/internal/usecase/recursos/ports/criteria"
	"github.com/evrone/go-clean-template/pkg/logger"
	"github.com/evrone/go-clean-template/pkg/roles"
	"github.com/gofiber/fiber/v2"
)

type RecursoRoutes struct {
	uc              recursosports.RecursoUseCase
	photoUseCase    recursosports.PhotoUseCase
	logErrorUseCase errorlogports.LogErrorUseCase
	logger          logger.Interface
}

func NewRecursoRoutes(router fiber.Router, uc recursosports.RecursoUseCase, photoUC recursosports.PhotoUseCase, logErrorUC errorlogports.LogErrorUseCase, l logger.Interface) {
	r := &RecursoRoutes{uc: uc, photoUseCase: photoUC, logErrorUseCase: logErrorUC, logger: l}
	h := router.Group("/recursos")
	{
		// Endpoints for resource administration (write operations)
		h.Post("/", middleware.RequireRole(roles.ADMIN_RECURSOS), r.createResource)
		h.Delete("/:id", middleware.RequireRole(roles.ADMIN_RECURSOS), r.deleteResource)
		h.Patch("/:id", middleware.RequireRole(roles.ADMIN_RECURSOS), r.updateResource)
		h.Post("/:id/foto", middleware.RequireRole(roles.ADMIN_RECURSOS), r.uploadPhoto)
		h.Delete("/:id/foto", middleware.RequireRole(roles.ADMIN_RECURSOS), r.deletePhoto)

		// Endpoints for general access (read operations)
		h.Get("/:id", r.getResourceByID)
		h.Get("/", r.listResources)
		h.Post("/search", r.searchResources)
	}
}

// --- Photo Handlers ---

// @Summary Subir foto de recurso
// @Description Sube una foto para un recurso específico. **Requiere rol:** ADMIN_RECURSOS
// @Tags recursos
// @Accept multipart/form-data
// @Produce json
// @Param id path int true "ID del recurso"
// @Param foto formData file true "Archivo de imagen"
// @Success 200 {object} entity.Recurso
// @Failure 400 {object} response.ErrorResponseDTO
// @Failure 403 {object} response.ErrorResponseDTO "Forbidden - Usuario sin permisos suficientes"
// @Failure 500 {object} response.ErrorResponseDTO
// @Router /recursos/{id}/foto [post]
// @Security BearerAuth
func (r *RecursoRoutes) uploadPhoto(c *fiber.Ctx) error {
	id, err := c.ParamsInt("id")
	if err != nil {
		r.logger.Warn("ID de recurso inválido en la URL para subir foto", "error", err.Error(), "path", c.Path())
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", "ID de recurso inválido", c.Path()))
	}

	file, err := c.FormFile("foto")
	if err != nil {
		r.logger.Warn("No se encontró el archivo 'foto' en la petición para subir foto", "error", err.Error(), "path", c.Path())
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", "No se encontró el archivo 'foto' en la petición", c.Path()))
	}

	input := DTOs.UploadPhotoInput{
		ResourceID: int64(id),
		File:       file,
	}

	recurso, err := r.photoUseCase.UploadPhoto(c.Context(), input)
	if err != nil {
		return r.handleError(c, err)
	}

	return c.Status(http.StatusOK).JSON(recurso)
}

// @Summary Eliminar foto de recurso
// @Description Elimina la foto de un recurso específico. **Requiere rol:** ADMIN_RECURSOS
// @Tags recursos
// @Produce json
// @Param id path int true "ID del recurso"
// @Success 204 "No Content"
// @Failure 400 {object} response.ErrorResponseDTO
// @Failure 403 {object} response.ErrorResponseDTO "Forbidden - Usuario sin permisos suficientes"
// @Failure 404 {object} response.ErrorResponseDTO
// @Failure 500 {object} response.ErrorResponseDTO
// @Router /recursos/{id}/foto [delete]
// @Security BearerAuth
func (r *RecursoRoutes) deletePhoto(c *fiber.Ctx) error {
	id, err := c.ParamsInt("id")
	if err != nil {
		r.logger.Warn("ID de recurso inválido en la URL para eliminar foto", "error", err.Error(), "path", c.Path())
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", "ID de recurso inválido", c.Path()))
	}

	err = r.photoUseCase.DeletePhoto(c.Context(), int64(id))
	if err != nil {
		return r.handleError(c, err)
	}

	return c.SendStatus(http.StatusNoContent)
}

// --- Handlers ---

// @Summary Listar recursos
// @Description Obtiene una lista paginada de recursos con filtros opcionales. **Acceso:** Cualquier usuario autenticado
// @Tags recursos
// @Accept json
// @Produce json
// @Param limit query int false "Número de elementos por página" default(10)
// @Param offset query int false "Desplazamiento para paginación" default(0)
// @Param nombre query string false "Filtrar por nombre"
// @Param estado query string false "Filtrar por estado (ACTIVO, MANTENIMIENTO)"
// @Success 200 {object} response.PaginatedResponse
// @Failure 400 {object} response.ErrorResponseDTO
// @Failure 500 {object} response.ErrorResponseDTO
// @Router /recursos [get]
// @Security BearerAuth
func (r *RecursoRoutes) listResources(c *fiber.Ctx) error {
	criteria, err := buildCriteriaFromQuery(c)
	if err != nil {
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", err.Error(), c.Path()))
	}

	paginatedOutput, err := r.uc.ListResources(c.Context(), criteria)
	if err != nil {
		return r.handleError(c, err)
	}

	response := buildPaginatedResponse(c, paginatedOutput)

	return c.Status(http.StatusOK).JSON(response)
}

// @Summary Buscar recursos
// @Description Busca recursos con criterios avanzados mediante POST. **Acceso:** Cualquier usuario autenticado
// @Tags recursos
// @Accept json
// @Produce json
// @Param limit query int false "Número de elementos por página" default(10)
// @Param offset query int false "Desplazamiento para paginación" default(0)
// @Param criteria body criteria.Criteria true "Criterios de búsqueda"
// @Success 200 {object} response.PaginatedResponse
// @Failure 400 {object} response.ErrorResponseDTO
// @Failure 500 {object} response.ErrorResponseDTO
// @Router /recursos/search [post]
// @Security BearerAuth
func (r *RecursoRoutes) searchResources(c *fiber.Ctx) error {
	var crit criteria.Criteria
	if err := c.BodyParser(&crit); err != nil {
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", "Cuerpo de la petición JSON inválido", c.Path()))
	}

	// Parse pagination from query params
	limit, _ := strconv.ParseUint(c.Query("limit", "10"), 10, 64)
	offset, _ := strconv.ParseUint(c.Query("offset", "0"), 10, 64)
	crit.Pagination = &criteria.Pagination{
		Limit:  limit,
		Offset: offset,
	}

	paginatedOutput, err := r.uc.ListResources(c.Context(), crit)
	if err != nil {
		return r.handleError(c, err)
	}

	response := buildPaginatedResponse(c, paginatedOutput)

	return c.Status(http.StatusOK).JSON(response)
}

// @Summary Crear recurso
// @Description Crea un nuevo recurso educativo. **Requiere rol:** ADMIN_RECURSOS
// @Tags recursos
// @Accept json
// @Produce json
// @Param recurso body request.CreateResource true "Datos del recurso a crear"
// @Success 201 {object} entity.Recurso
// @Failure 400 {object} response.ErrorResponseDTO
// @Failure 403 {object} response.ErrorResponseDTO "Forbidden - Usuario sin permisos suficientes"
// @Failure 500 {object} response.ErrorResponseDTO
// @Router /recursos [post]
// @Security BearerAuth
func (r *RecursoRoutes) createResource(c *fiber.Ctx) error {
	var req request.CreateResource
	if err := c.BodyParser(&req); err != nil {
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", "Cuerpo de la petición inválido", c.Path()))
	}

	input := toCreateResourceInput(req)

	recurso, err := r.uc.CreateResource(c.Context(), input)
	if err != nil {
		return r.handleError(c, err)
	}

	return c.Status(http.StatusCreated).JSON(recurso)
}

// @Summary Eliminar recurso
// @Description Elimina un recurso por su ID. **Requiere rol:** ADMIN_RECURSOS
// @Tags recursos
// @Produce json
// @Param id path int true "ID del recurso"
// @Success 204 "No Content"
// @Failure 400 {object} response.ErrorResponseDTO
// @Failure 403 {object} response.ErrorResponseDTO "Forbidden - Usuario sin permisos suficientes"
// @Failure 404 {object} response.ErrorResponseDTO
// @Failure 500 {object} response.ErrorResponseDTO
// @Router /recursos/{id} [delete]
// @Security BearerAuth
func (r *RecursoRoutes) deleteResource(c *fiber.Ctx) error {
	id, err := c.ParamsInt("id")
	if err != nil {
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", "ID de recurso inválido", c.Path()))
	}

	err = r.uc.DeleteResource(c.Context(), int64(id))
	if err != nil {
		return r.handleError(c, err)
	}

	return c.SendStatus(http.StatusNoContent)
}

// @Summary Actualizar recurso
// @Description Actualiza parcialmente un recurso por su ID. **Requiere rol:** ADMIN_RECURSOS
// @Tags recursos
// @Accept json
// @Produce json
// @Param id path int true "ID del recurso"
// @Param recurso body request.UpdateResource true "Datos del recurso a actualizar"
// @Success 200 {object} entity.Recurso
// @Failure 400 {object} response.ErrorResponseDTO
// @Failure 403 {object} response.ErrorResponseDTO "Forbidden - Usuario sin permisos suficientes"
// @Failure 404 {object} response.ErrorResponseDTO
// @Failure 500 {object} response.ErrorResponseDTO
// @Router /recursos/{id} [patch]
// @Security BearerAuth
func (r *RecursoRoutes) updateResource(c *fiber.Ctx) error {
	id, err := c.ParamsInt("id")
	if err != nil {
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", "ID de recurso inválido", c.Path()))
	}

	var req request.UpdateResource
	if err := c.BodyParser(&req); err != nil {
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", "Cuerpo de la petición inválido", c.Path()))
	}

	input := toUpdateResourceInput(req)

	recurso, err := r.uc.UpdateResource(c.Context(), int64(id), input)
	if err != nil {
		return r.handleError(c, err)
	}

	return c.Status(http.StatusOK).JSON(recurso)
}

// @Summary Obtener recurso por ID
// @Description Obtiene los detalles de un recurso específico por su ID. **Acceso:** Cualquier usuario autenticado
// @Tags recursos
// @Produce json
// @Param id path int true "ID del recurso"
// @Success 200 {object} entity.Recurso
// @Failure 400 {object} response.ErrorResponseDTO
// @Failure 404 {object} response.ErrorResponseDTO
// @Failure 500 {object} response.ErrorResponseDTO
// @Router /recursos/{id} [get]
// @Security BearerAuth
func (r *RecursoRoutes) getResourceByID(c *fiber.Ctx) error {
	id, err := c.ParamsInt("id")
	if err != nil {
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", "ID de recurso inválido", c.Path()))
	}

	recurso, err := r.uc.GetResourceByID(c.Context(), int64(id))
	if err != nil {
		return r.handleError(c, err)
	}

	return c.Status(http.StatusOK).JSON(recurso)
}

// --- Mappers & Helpers ---

func toCreateResourceInput(req request.CreateResource) DTOs.CreateResourceInput {
	return DTOs.CreateResourceInput{
		Nombre:      req.Nombre,
		Descripcion: req.Descripcion,
		HrefPhoto:   req.HrefPhoto,
		Estado:      req.Estado,
	}
}

func toUpdateResourceInput(req request.UpdateResource) DTOs.UpdateResourceInput {
	return DTOs.UpdateResourceInput{
		Nombre:      req.Nombre,
		Descripcion: req.Descripcion,
		HrefPhoto:   req.HrefPhoto,
		Estado:      req.Estado,
	}
}

func (r *RecursoRoutes) handleError(c *fiber.Ctx, err error) error {
	switch {
	case errors.Is(err, apperror.ErrBadRequest):
		r.logger.Warn("Bad Request", "error", err.Error(), "path", c.Path())
		return c.Status(http.StatusBadRequest).JSON(common_response.NewErrorResponseDTO(http.StatusBadRequest, "Bad Request", err.Error(), c.Path()))
	case errors.Is(err, apperror.ErrConflict):
		r.logger.Warn("Conflict error", "error", err.Error(), "path", c.Path())
		return c.Status(http.StatusConflict).JSON(common_response.NewErrorResponseDTO(http.StatusConflict, "Conflict", err.Error(), c.Path()))
	case errors.Is(err, apperror.ErrNotFound):
		r.logger.Warn("Not Found error", "error", err.Error(), "path", c.Path())
		return c.Status(http.StatusNotFound).JSON(common_response.NewErrorResponseDTO(http.StatusNotFound, "Not Found", err.Error(), c.Path()))
	default:
		r.logger.Error("Internal Server Error", "error", err.Error(), "path", c.Path())

		if loggingErr := r.logErrorUseCase.Execute(c.Context(), errorlogports.LogErrorInput{
			Err:           err,
			RequestPath:   c.Path(),
			RequestMethod: c.Method(),
		}); loggingErr != nil {
			r.logger.Error("Fallo al persistir el error interno en la BD", "error_de_logging", loggingErr.Error())
		}

		return c.Status(http.StatusInternalServerError).JSON(common_response.NewErrorResponseDTO(http.StatusInternalServerError, "Internal Server Error", "Ocurrió un error inesperado", c.Path()))
	}
}

// buildCriteriaFromQuery parsea los query params de la URL para construir un objeto Criteria.
// Ejemplo: ?filter[nombre][CONTAINS]=laptop&filter[estado][EQUALS]=ACTIVO&sort=-created_at&page=1&limit=10&logical_op=OR
func buildCriteriaFromQuery(c *fiber.Ctx) (criteria.Criteria, error) {
	var filters []criteria.Filter
	for key, value := range c.Queries() {
		if strings.HasPrefix(key, "filter[") {
			parts := strings.Split(strings.TrimSuffix(key, "]"), "[")
			if len(parts) == 4 { // filter, field, operator, ""
				field := parts[1]
				opStr := strings.TrimSuffix(parts[2], "]")
				filters = append(filters, criteria.Filter{
					Field:    field,
					Operator: criteria.FilterOperator(opStr),
					Value:    value,
				})
			}
		}
	}

	var sorts []criteria.Sort
	if sortQuery := c.Query("sort"); sortQuery != "" {
		fields := strings.Split(sortQuery, ",")
		for _, f := range fields {
			direction := criteria.ASC
			if strings.HasPrefix(f, "-") {
				direction = criteria.DESC
				f = strings.TrimPrefix(f, "-")
			}
			sorts = append(sorts, criteria.Sort{Field: f, Direction: direction})
		}
	}

	var pag *criteria.Pagination
	page, _ := strconv.ParseUint(c.Query("page", "1"), 10, 64)
	limit, _ := strconv.ParseUint(c.Query("limit", "10"), 10, 64)
	if page > 0 {
		pag = &criteria.Pagination{
			Limit:  limit,
			Offset: (page - 1) * limit,
		}
	}

	logicalOp := criteria.AND
	if strings.ToUpper(c.Query("logical_op")) == "OR" {
		logicalOp = criteria.OR
	}

	return criteria.Criteria{
		FilterGroup: criteria.FilterGroup{
			Operator: logicalOp,
			Filters:  filters,
		},
		Sort:       sorts,
		Pagination: pag,
	}, nil
}

func buildPaginatedResponse(c *fiber.Ctx, paginatedOutput *DTOs.PaginatedRecursosOutput) response.PaginatedResponse {
	// 1. Map entities to HATEOAS response DTOs
	recursosHATEOAS := make([]response.RecursoHATEOAS, 0, len(paginatedOutput.Recursos))
	for _, r := range paginatedOutput.Recursos {
		recursosHATEOAS = append(recursosHATEOAS, response.RecursoHATEOAS{
			RecursoResponse: response.RecursoResponse{
				ID:          r.ID,
				Nombre:      r.Nombre,
				Descripcion: r.Descripcion,
				HrefPhoto:   r.HrefPhoto,
				Estado:      r.Estado,
				CreatedAt:   r.CreatedAt,
				UpdatedAt:   r.UpdatedAt,
			},
			Links: struct {
				Self response.Link `json:"self"`
			}{
				Self: response.Link{
					Href: fmt.Sprintf("%s/recursos/%d", c.BaseURL(), r.ID),
				},
			},
		})
	}

	// 2. Calculate page info
	totalPages := int64(0)
	if paginatedOutput.Limit > 0 {
		totalPages = int64(math.Ceil(float64(paginatedOutput.TotalElements) / float64(paginatedOutput.Limit)))
	}
	currentPage := int64(0)
	if paginatedOutput.Limit > 0 {
		currentPage = int64(paginatedOutput.Offset / paginatedOutput.Limit)
	}

	pageInfo := response.PageInfo{
		Size:          uint64(len(paginatedOutput.Recursos)),
		TotalElements: paginatedOutput.TotalElements,
		TotalPages:    totalPages,
		Number:        currentPage,
	}

	// 3. Build navigation links
	selfURL := c.OriginalURL()
	var nextURL, prevURL *response.Link

	// Next page link
	if (paginatedOutput.Offset + paginatedOutput.Limit) < uint64(paginatedOutput.TotalElements) {
		nextOffset := paginatedOutput.Offset + paginatedOutput.Limit
		nextURL = &response.Link{
			Href: fmt.Sprintf("%s/recursos/search?limit=%d&offset=%d", c.BaseURL(), paginatedOutput.Limit, nextOffset),
		}
	}

	// Previous page link
	if paginatedOutput.Offset > 0 {
		prevOffset := paginatedOutput.Offset - paginatedOutput.Limit
		if int64(prevOffset) < 0 { // Ensure offset doesn't go negative
			prevOffset = 0
		}
		prevURL = &response.Link{
			Href: fmt.Sprintf("%s/recursos/search?limit=%d&offset=%d", c.BaseURL(), paginatedOutput.Limit, prevOffset),
		}
	}

	links := response.Links{
		Self: response.Link{Href: selfURL},
		Next: nextURL,
		Prev: prevURL,
	}

	// 4. Assemble final response
	return response.PaginatedResponse{
		Embedded: response.EmbeddedRecursos{
			Recursos: recursosHATEOAS,
		},
		Links: links,
		Page:  pageInfo,
	}
}